@@ -1,1 +1,1 @@
-module.exports = require('./eslint.strapi.json')
+module.exports = require('./eslint.json')
