@@ -1,10 +1,24 @@
-import { RuleSetRule } from 'webpack'
+import { RuleSetRule, RuleSetUseItem } from 'webpack'
 import MiniCssExtractPlugin from 'mini-css-extract-plugin'
 import path from 'path'
-import autoprefixer from 'autoprefixer'
-export const babel: RuleSetRule = {
-  test: /\.(js|ts)x?$/i,
+
+export const typescript: RuleSetRule = {
+  test: /\.(js|ts)$/i,
   loader: 'babel-loader',
+  exclude: /node_modules/,
+  options: {
+    exclude: [/node_modules[\\/]core-js/, /node_modules[\\/]webpack[\\/]buildin/],
+    cacheDirectory: true,
+    cacheCompression: false,
+    compact: process.env.NODE_ENV === 'production',
+    presets: [['@babel/preset-env'], ['@babel/preset-typescript']],
+  },
+}
+
+export const react: RuleSetRule = {
+  test: /\.(js|mjs|jsx|ts|tsx)$/,
+  loader: require.resolve('babel-loader'),
+  include: /src/,
   exclude: /node_modules/,
   options: {
     exclude: [/node_modules[\\/]core-js/, /node_modules[\\/]webpack[\\/]buildin/],
@@ -25,7 +39,7 @@ export const babel: RuleSetRule = {
 }
 
 export const ejs: RuleSetRule = {
-  test: /\.(ejs)$/i,
+  test: /\.(ejs)$/,
   loader: 'ejs-loader',
   exclude: /node-modules/,
   options: {
@@ -33,32 +47,8 @@ export const ejs: RuleSetRule = {
   },
 }
 
-export const less: RuleSetRule = {
-  test: /\.((c|sa|sc|le)ss)$/i,
-  use: [
-    {
-      loader: MiniCssExtractPlugin.loader,
-    },
-    {
-      loader: require.resolve('css-loader'),
-      options: { sourceMap: true },
-    },
-    {
-      loader: require.resolve('less-loader'),
-      options: {
-        sourceMap: true,
-        implementation: require.resolve('less'),
-        lessOptions: {
-          paths: [path.join(process.cwd()), path.join(process.cwd(), 'node_modules')],
-          javascriptEnabled: true,
-        },
-      },
-    },
-  ],
-}
-
-export const lessPostCss: RuleSetRule = {
-  test: /\.((c|sa|sc|le)ss)$/i,
+const styles: RuleSetRule = {
+  test: /\.(css)$/,
   use: [
     MiniCssExtractPlugin.loader,
     {
@@ -69,12 +59,18 @@ export const lessPostCss: RuleSetRule = {
       loader: require.resolve('postcss-loader'),
       options: {
         sourceMap: true,
-        implementation: require.resolve('postcss'),
         postcssOptions: {
-          plugins: [autoprefixer],
+          ident: 'postcss',
+          plugins: [['postcss-preset-env']],
         },
       },
     },
+  ],
+}
+
+export const less: RuleSetRule = {
+  test: /\.(css|less)$/,
+  use: (styles.use as Array<RuleSetUseItem>).concat([
     {
       loader: require.resolve('less-loader'),
       options: {
@@ -86,44 +82,68 @@ export const lessPostCss: RuleSetRule = {
         },
       },
     },
-  ],
+  ]),
+}
+
+export const sass: RuleSetRule = {
+  test: /\.(css|scss|sass)$/,
+  use: (styles.use as Array<RuleSetUseItem>).concat([
+    {
+      loader: require.resolve('sass-loader'),
+      options: {
+        sourceMap: true,
+        implementation: require.resolve('sass'),
+      },
+    },
+  ]),
 }
 
 export const fonts: RuleSetRule = {
-  test: /\.(woff|woff2|eot|ttf|otf)$/i,
-  type: 'asset/resource',
-}
-
-export const htmlLoader: RuleSetRule = {
-  test: /\.html$/i,
-  loader: 'html-loader',
-  options: {
-    minimize: true,
+  test: /\.(woff|woff2|otf|ttf|eot)$/,
+  type: 'asset',
+  parser: {
+    dataUrlCondition: {
+      maxSize: 0,
+    },
   },
 }
 
+export const html: RuleSetRule = {
+  test: /\.html$/,
+  loader: require.resolve('html-loader'),
+}
+
 export const images: RuleSetRule = {
-  test: /\.(png|webp|bmp|jpg|jpeg|gif)$/i,
+  test: [/\.bmp$/, /\.gif$/, /\.webp$/, /\.jpe?g$/, /\.png$/],
   type: 'asset',
+  parser: {
+    dataUrlCondition: {
+      maxSize: 1024,
+    },
+  },
+}
+export const svgUrl: RuleSetRule = {
+  test: /\.svg$/,
+  loader: require.resolve('url-loader'),
+  options: {
+    limit: 8192,
+  },
+  issuer: /\.(c|sc|le|sa)ss$/,
 }
 
-export const svgInStyles: RuleSetRule = {
-  test: /\.svg$/i,
-  loader: 'url-loader',
+export const videos: RuleSetRule = {
+  test: /\.(mp4|mov|flv)$/,
+  type: 'asset/resource',
+  parser: {
+    dataUrlCondition: {
+      maxSize: 0,
+    },
+  },
 }
 
-export const svgInComponents: RuleSetRule = {
-  test: /\.svg$/i,
-  issuer: /\.(js|ts)x?$/,
-  use: [
-    {
-      loader: '@svgr/webpack',
-      options: {
-        babel: false,
-      },
-    },
-    {
-      loader: 'url-loader',
-    },
-  ],
+export const svgComponent: RuleSetRule = {
+  test: /\.svg$/,
+  exclude: /node_modules/,
+  use: [{ loader: require.resolve('@svgr/webpack') }, { loader: require.resolve('url-loader') }],
+  issuer: /\.(js|ts)x$/,
 }
